Separating mixer from its nodes and clippy fixes
- Mixer operations are now separate from the node operations.
- FFT operations were moved into src/fft/mod.rs to avoid redundancy in
  imports and to fix clippy issues.
